+++ src/athi_utility.cpp
@@ -47,13 +47,13 @@ vec3 rand_vec3(f32 min, f32 max) noexcept { return vec3(rand_f32(min, max), rand
 vec4 rand_vec4(f32 min, f32 max) noexcept { return vec4(rand_f32(min, max), rand_f32(min, max), rand_f32(min, max), rand_f32(min, max)); }
 
 
-std::tuple<size_t,size_t> get_begin_and_end(s32 i, s32 total, s32 threads) noexcept {
+std::tuple<s32,s32> get_begin_and_end(s32 i, s32 total, s32 threads) noexcept {
   const s32 parts = total / threads;
   const s32 leftovers = total % threads;
   const s32 begin = parts * i;
   s32 end = parts * (i + 1);
   if (i == threads - 1) end += leftovers;
-  return std::tuple<size_t, size_t>{begin, end};
+  return std::tuple<s32, s32>{begin, end};
 }
